+++ ext/kernel/memory.c
@@ -191,7 +191,7 @@ void phalcon_dump_current_frame(TSRMLS_D)
 {
 	zend_phalcon_globals *phalcon_globals_ptr = PHALCON_VGLOBAL;
 	phalcon_memory_entry *active_memory;
-	int i;
+	size_t i;
 
 	if (unlikely(phalcon_globals_ptr->active_memory == NULL)) {
 		fprintf(stderr, "WARNING: calling phalcon_dump_current_frame() without an active memory frame!\n");
@@ -207,14 +207,14 @@ void phalcon_dump_current_frame(TSRMLS_D)
 	if (active_memory->hash_pointer) {
 		for (i = 0; i < active_memory->hash_pointer; ++i) {
 			assert(active_memory->hash_addresses[i] != NULL && *(active_memory->hash_addresses[i]) != NULL);
-			fprintf(stderr, "Hash ptr %d (%p => %p), type=%u, refcnt=%u\n", i, active_memory->hash_addresses[i], *active_memory->hash_addresses[i], Z_TYPE_PP(active_memory->hash_addresses[i]), Z_REFCOUNT_PP(active_memory->hash_addresses[i]));
+			fprintf(stderr, "Hash ptr %lu (%p => %p), type=%u, refcnt=%u\n", (ulong)i, active_memory->hash_addresses[i], *active_memory->hash_addresses[i], Z_TYPE_PP(active_memory->hash_addresses[i]), Z_REFCOUNT_PP(active_memory->hash_addresses[i]));
 		}
 	}
 
 	for (i = 0; i < active_memory->pointer; ++i) {
 		if (likely(active_memory->addresses[i] != NULL && *(active_memory->addresses[i]) != NULL)) {
 			zval **var = active_memory->addresses[i];
-			fprintf(stderr, "Obs var %d (%p => %p), type=%u, refcnt=%u; ", i, var, *var, Z_TYPE_PP(var), Z_REFCOUNT_PP(var));
+			fprintf(stderr, "Obs var %lu (%p => %p), type=%u, refcnt=%u; ", (ulong)i, var, *var, Z_TYPE_PP(var), Z_REFCOUNT_PP(var));
 			switch (Z_TYPE_PP(var)) {
 				case IS_NULL:     fprintf(stderr, "value=NULL\n"); break;
 				case IS_LONG:     fprintf(stderr, "value=%ld\n", Z_LVAL_PP(var)); break;
